updates from PR review
